# backend/docs/swagger.yaml
definitions:
  database.Announcement:
    properties:
      announcement_created_at:
        type: string
      announcement_created_by:
        type: integer
      announcement_id:
        type: integer
      announcement_message:
        type: string
      announcement_title:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /announcements:
    get:
      description: Retrieves all announcements in descending order by creation date
      produces:
      - application/json
      responses:
        "200":
          description: List of announcements
          schema:
            items:
              $ref: '#/definitions/database.Announcement'
            type: array
        "500":
          description: Error retrieving announcements
          schema:
            type: string
      summary: Retrieve all announcements
      tags:
      - Announcements
  /announcements/create:
    post:
      consumes:
      - application/json
      description: Allows an admin user to create a new announcement
      parameters:
      - description: Announcement data
        in: body
        name: announcement
        required: true
        schema:
          $ref: '#/definitions/database.Announcement'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created announcement
          schema:
            $ref: '#/definitions/database.Announcement'
        "400":
          description: Invalid request body or missing required fields
          schema:
            type: string
        "401":
          description: Only admins can post announcements
          schema:
            type: string
        "500":
          description: Error saving announcement
          schema:
            type: string
      summary: Create a new announcement
      tags:
      - Announcements
...
swagger: "2.0"